+++ to-do-app/src/components/ListItem/index.js
@@ -26,13 +26,11 @@ class ListItem extends Component {
   render() {
     return (
         <div className="item-list">
-            <div className = "List-Item-Button">
-              <Checkbox 
-                defaultChecked={this.props.itemProperties.completed}
-                onChange={this.handleClick}/>
-            </div>
-            <div className="item">{this.props.itemProperties.toDoItem}</div>
-            <img src="/remove-item.png" 
+          <Checkbox 
+            defaultChecked={this.props.itemProperties.completed}
+            onChange={this.handleClick}/>
+          <div className="item">{this.props.itemProperties.toDoItem}</div>
+            <img src="/remove-item.png"
             className="Action-Button"
             onClick={this.handleDelete}></img>
         </div>
