+++ to-do-app/src/components/NewItem/index.test.js
@@ -1,10 +1,10 @@
 import React from 'react';
-import ActionForm from '.';
+import NewItem from '.';
 import { render, screen, fireEvent, userEvent } from "@testing-library/react"
 import { shallow, mount } from "enzyme";
 
 test("renders add button image", () => {
-  render(<ActionForm />);
+  render(<NewItem />);
 
   const image = screen.getByRole("img");
 
@@ -12,7 +12,7 @@ test("renders add button image", () => {
 });
 
 test("renders form for a new item", () => {
-  render(<ActionForm />);
+  render(<NewItem />);
 
   const form = screen.getAllByRole("textbox");
 
@@ -27,9 +27,9 @@ test("adds a new item to the active list", () => {
   const testList = { "id": 1, "title": "Dummy Title", "items": 
                       [ {"toDoItem": "First Item", "completed": false} ] }
 
-  const wrapper = shallow(<ActionForm activeList={testList} saveActiveList={() => null}/>);
+  const wrapper = shallow(<NewItem activeList={testList} saveActiveList={() => null}/>);
   
-  render(<ActionForm activeList={testList} saveActiveList={() => null}/>);
+  render(<NewItem activeList={testList} saveActiveList={() => null}/>);
 
   const inputForm = wrapper.find("input");
